refactor(MyKSuiteDashboard): Extract composable to their own files for better readibility
