+++ MykSuite/src/main/java/com/infomaniak/lib/myksuite/ui/screens/MyKSuiteDashboardScreen.kt
@@ -18,41 +18,28 @@
 package com.infomaniak.lib.myksuite.ui.screens
 
 import android.content.res.Configuration
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.border
-import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.annotation.DrawableRes
-import androidx.annotation.StringRes
-import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.*
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.*
-import androidx.compose.runtime.*
-import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.StrokeCap
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import coil3.compose.AsyncImage
-import coil3.request.ImageRequest
-import coil3.request.crossfade
 import com.infomaniak.lib.myksuite.R
 import com.infomaniak.lib.myksuite.ui.components.MyKSuiteChip
-import com.infomaniak.lib.myksuite.ui.components.WeightOneSpacer
+import com.infomaniak.lib.myksuite.ui.screens.components.AppStorageQuotas
 import com.infomaniak.lib.myksuite.ui.screens.components.ExpendableActionItem
+import com.infomaniak.lib.myksuite.ui.screens.components.LimitedFunctionalities
+import com.infomaniak.lib.myksuite.ui.screens.components.UserAvatar
 import com.infomaniak.lib.myksuite.ui.theme.Dimens
 import com.infomaniak.lib.myksuite.ui.theme.Margin
 import com.infomaniak.lib.myksuite.ui.theme.MyKSuiteTheme
@@ -96,52 +83,14 @@ fun MyKSuiteDashboardScreen(
                 PaddedDivider(paddedModifier)
                 AppStorageQuotas(paddedModifier)
                 PaddedDivider(paddedModifier)
+                ExpendableActionItem(iconRes = R.drawable.ic_envelope, textRes = R.string.myKSuiteDashboardFreeMailLabel)
+                ExpendableActionItem(
+                    iconRes = R.drawable.ic_padlock,
+                    textRes = R.string.myKSuiteDashboardLimitedFunctionalityLabel,
+                    expendedView = { LimitedFunctionalities(paddedModifier, dailySendingLimit) },
+                )
+                Spacer(Modifier.height(Margin.Medium))
             }
-        }
-    }
-}
-
-@Composable
-private fun UserAvatar(avatarUri: String) {
-    val context = LocalContext.current
-    var shouldDisplayPreview by rememberSaveable(avatarUri) { mutableStateOf(true) }
-
-    Box {
-        if (shouldDisplayPreview) {
-            val imageRequest = remember(avatarUri) {
-                ImageRequest.Builder(context)
-                    .data(avatarUri)
-                    .crossfade(true)
-                    .build()
-            }
-
-            AsyncImage(
-                modifier = Modifier.size(Dimens.iconSize),
-                model = imageRequest,
-                contentDescription = null,
-                contentScale = ContentScale.Crop,
-                onError = { shouldDisplayPreview = false },
-            )
-        } else {
-            Image(
-                modifier = Modifier
-                    .size(Dimens.iconSize)
-                    .border(
-                        width = 1.dp,
-                        brush = Brush.linearGradient(
-                            0.0f to Color(0xFF1DDDFD),
-                            0.1f to Color(0xFF337CFF),
-                            0.3f to Color(0xFFA055FC),
-                            0.4f to Color(0xFFF34BBB),
-                            0.6f to Color(0xFFFD8C3D),
-                            start = Offset(0.0f, 10.0f),
-                            end = Offset(0.0f, 90.0f),
-                        ),
-                        shape = RoundedCornerShape(Dimens.largeCornerRadius),
-                    ),
-                imageVector = ImageVector.vectorResource(R.drawable.ic_person),
-                contentDescription = null,
-            )
         }
     }
 }
@@ -180,48 +129,6 @@ private fun PaddedDivider(modifier: Modifier) {
     HorizontalDivider(modifier)
     Spacer(Modifier.height(Margin.Large))
 }
-
-@Composable
-private fun AppStorageQuotas(modifier: Modifier) {
-    Column(modifier = modifier) {
-        KSuiteApp.entries.forEach {
-            AppStorageQuota(app = it)
-            if (it.ordinal != KSuiteApp.entries.lastIndex) Spacer(Modifier.height(Margin.Medium))
-        }
-    }
-}
-
-@Composable
-private fun AppStorageQuota(modifier: Modifier = Modifier, app: KSuiteApp) {
-    Column(modifier) {
-        Row(verticalAlignment = Alignment.CenterVertically) {
-            Text(app.displayName)
-            WeightOneSpacer(minWidth = Margin.Medium)
-            Text(
-                "0.2 Go / 20 Go", // TODO: Use real data
-                style = MyKSuiteTheme.typography.bodySmallRegular,
-                color = MyKSuiteTheme.colors.secondaryTextColor,
-            )
-        }
-        Spacer(Modifier.height(Margin.Mini))
-        LinearProgressIndicator(
-            modifier = Modifier
-                .height(14.dp)
-                .fillMaxWidth(),
-            color = app.color(),
-            trackColor = MyKSuiteTheme.colors.chipBackground,
-            strokeCap = StrokeCap.Round,
-            gapSize = -Margin.Mini,
-            progress = { 0.5f }, // TODO: Use real values
-            drawStopIndicator = {},
-        )
-    }
-}
-
-private enum class KSuiteApp(val displayName: String, val color: @Composable () -> Color) {
-    Mail("Mail", { MyKSuiteTheme.colors.mail }), Drive("kDrive", { MyKSuiteTheme.colors.drive })
-}
-
 @Preview(name = "(1) Light")
 @Preview(name = "(2) Dark", uiMode = Configuration.UI_MODE_NIGHT_YES or Configuration.UI_MODE_TYPE_NORMAL)
 @Composable
